Add EclipseLink and Declarative service JPA examples
