+++ examples/karaf-jpa-example/karaf-jpa-example-provider/karaf-jpa-example-provider-api/src/main/java/org/apache/karaf/examples/jpa/BookingService.java
