+++ examples/karaf-jpa-example/karaf-jpa-example-provider/karaf-jpa-example-provider-blueprint/karaf-jpa-example-provider-blueprint-hibernate/pom.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <!--
 
@@ -19,17 +21,15 @@
         limitations under the License.
     -->
 
+    <parent>
+        <artifactId>karaf-jpa-example-provider-blueprint</artifactId>
+        <groupId>org.apache.karaf.examples</groupId>
+        <version>4.2.1-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <parent>
-        <groupId>org.apache.karaf.examples</groupId>
-        <artifactId>karaf-jpa-example</artifactId>
-        <version>4.2.1-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>karaf-jpa-example-command</artifactId>
-    <name>Apache Karaf :: Examples :: JPA :: Command</name>
+    <artifactId>karaf-jpa-example-provider-blueprint-hibernate</artifactId>
+    <name>Apache Karaf :: Examples :: JPA :: Provider :: Blueprint :: Hibernate</name>
     <packaging>bundle</packaging>
 
     <dependencies>
@@ -39,29 +39,32 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.karaf.shell</groupId>
-            <artifactId>org.apache.karaf.shell.core</artifactId>
+            <groupId>org.eclipse.persistence</groupId>
+            <artifactId>javax.persistence</artifactId>
+            <version>2.1.0</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.transaction</groupId>
+            <artifactId>javax.transaction-api</artifactId>
+            <version>1.2</version>
+        </dependency>
+        <dependency>
+            <groupId>javax.interceptor</groupId>
+            <artifactId>javax.interceptor-api</artifactId>
+            <version>1.2</version>
         </dependency>
     </dependencies>
 
     <build>
         <plugins>
             <plugin>
-                <groupId>org.apache.karaf.tooling</groupId>
-                <artifactId>karaf-services-maven-plugin</artifactId>
-            </plugin>
-            <plugin>
                 <groupId>org.apache.felix</groupId>
                 <artifactId>maven-bundle-plugin</artifactId>
                 <configuration>
                     <instructions>
-                        <Private-Package>
-                            org.apache.karaf.examples.jpa.command,
-                            org.apache.karaf.examples.jpa.completers
-                        </Private-Package>
+                        <Meta-Persistence>META-INF/persistence.xml</Meta-Persistence>
                         <Import-Package>
-                            org.apache.karaf.examples.jpa,
-                            org.apache.karaf.shell*;version="[4,5)",
+                            javax.persistence;version="[2,3)",
                             *
                         </Import-Package>
                     </instructions>
@@ -69,5 +72,4 @@
             </plugin>
         </plugins>
     </build>
-
 </project>
