+++ examples/karaf-jpa-example/karaf-jpa-example-provider/karaf-jpa-example-provider-ds/pom.xml
@@ -1,5 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 
     <!--
 
@@ -19,23 +21,21 @@
         limitations under the License.
     -->
 
+    <parent>
+        <artifactId>karaf-jpa-example-provider</artifactId>
+        <groupId>org.apache.karaf.examples</groupId>
+        <version>4.2.1-SNAPSHOT</version>
+    </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <parent>
-        <groupId>org.apache.karaf.examples</groupId>
-        <artifactId>karaf-jpa-example</artifactId>
-        <version>4.2.1-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-
-    <artifactId>karaf-jpa-example-provider</artifactId>
-    <name>Apache Karaf :: Examples :: JPA :: Provider</name>
+    <artifactId>karaf-jpa-example-provider-ds</artifactId>
+    <name>Apache Karaf :: Examples :: JPA :: Provider :: Declarative Service</name>
     <packaging>pom</packaging>
 
     <modules>
-        <module>karaf-jpa-example-provider-api</module>
-        <module>karaf-jpa-example-provider-ds</module>
-        <module>karaf-jpa-example-provider-blueprint</module>
+        <module>karaf-jpa-example-provider-ds-eclipselink</module>
+        <module>karaf-jpa-example-provider-ds-hibernate</module>
+        <module>karaf-jpa-example-provider-ds-openjpa</module>
     </modules>
 
 </project>
