+++ itests/test/pom.xml
@@ -231,7 +231,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.karaf.examples</groupId>
-            <artifactId>karaf-jpa-example-provider</artifactId>
+            <artifactId>karaf-jpa-example-provider-api</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
